==> Notebooks/1_Loading_Data.ipynb <==
import pandas as pd

# Read in and examine the imdb reviews dataset
imdb = pd.read_csv('CSVs/IMDB Dataset.csv')
imdb.head()

imdb['sentiment'].unique()

# Change sentiment to match format of other two datasets--2 for positive, 1 for negative
imdb.sentiment = imdb.sentiment.apply(lambda x: 2 if x=='positive' else 1)

# Put the columns in order
imdb = imdb[['sentiment','review']]
imdb.head()

# Read in and examine the yelp reviews dataset
yelp_train = pd.read_csv('CSVs/yelp_train.csv',header=None)
yelp_train.head()

yelp_train.columns = ['sentiment','review']
yelp_train.head()

yelp_test = pd.read_csv('CSVs/yelp_test.csv',header=None)
yelp_test.head()

yelp_test.columns  = ['sentiment','review']
yelp_test.head()

def preprocess_amazon(input_file, output_file, encoding):
    '''Amazon files are in the format __label__[1,2] [text].
    Create a processed file in the format [1,2]\t[text]'''
    with open(input_file, 'r',encoding=file_encoding) as infile, open(output_file, 'w',encoding=file_encoding) as outfile:
        i = 0
        for line in infile: 
            processed_line = line.strip('__label__')
            processed_line = processed_line[0] + '\t' + processed_line[2:]
            outfile.write(processed_line + '\n')


test_input_file = 'CSVs/amazon_test.ft.txt'
test_output_file = 'CSVs/amazon_test_processed.txt'
train_input_file = 'CSVs/amazon_train.ft.txt'
train_output_file = 'CSVs/amazon_train_processed.txt'
file_encoding = 'utf-8'

# Create new, tab-delimited files
preprocess_amazon(test_input_file,test_output_file,file_encoding)

preprocess_amazon(train_input_file,train_output_file,file_encoding)

amazon_test = pd.read_csv('CSVs/amazon_test_processed.txt',sep='\t',header=None,names=['sentiment','review'])
amazon_test.head()

amazon_train = pd.read_csv('CSVs/amazon_train_processed.txt',sep='\t',header=None,names=['sentiment','review'])
amazon_train.head()

# Recombine data that was split in original dataset
yelp = pd.concat([yelp_train,yelp_test],ignore_index=True)
len(yelp)

amazon = pd.concat([amazon_train,amazon_test],ignore_index=True)
len(amazon)

len(imdb)

# Label data sources for stratification
imdb['source'] = 'imdb'
yelp['source'] = 'yelp'
amazon['source'] = 'amazon'

reviews = pd.concat([imdb, yelp, amazon],ignore_index=True)

reviews['sentiment'].unique()

from sklearn.model_selection import train_test_split

# Split combined data into train/test, stratified by data source
X_train, X_test, y_train, y_test = train_test_split(reviews['review'],reviews['sentiment'],test_size=0.1,stratify=reviews['source'],random_state=42)

print(len(y_train),len(y_test))

y_train.head()

X_train.head()

# Save split data for later use
X_train.to_csv('CSVs/X_train_full.csv')
X_test.to_csv('CSVs/X_test_full.csv')
y_train.to_csv('CSVs/y_train_full.csv')
y_test.to_csv('CSVs/y_test_full.csv')
