==> Notebooks/3_Preprocessing.ipynb <==
!pip install transformers

from google.colab import drive
drive.mount('/content/drive/')

import pandas as pd
import dask.dataframe as dd
from transformers import AutoTokenizer
import pyarrow as pa

#X_train = pd.read_csv('drive/MyDrive/BERT Sentiment/CSVs/X_train_full.csv')
X_test = pd.read_csv('drive/MyDrive/BERT Sentiment/CSVs/X_test_full.csv')

#X_train.columns = ['index','review']
X_test.columns = ['index','review']

#X_train.set_index(keys='index',inplace=True)
X_test.set_index('index',inplace=True)
#X_train.head()
X_test.head()

# Load HuggingFace Tokenizer
tokenizer = AutoTokenizer.from_pretrained('distilbert-base-uncased')

# Tokenize reviews
test_inputs = tokenizer(list(X_test['review']),truncation=True,padding=True)

test_inputs.keys()


# Examining reviews with unknown tokens
has_unknown = []
for i, input in enumerate(test_inputs['input_ids']):
  for tok in input:
    # Searches for an [UNK] token
    if tok == 100:
      # Prints entire line with padding characters removed
      print(f"input {i}: {tokenizer.decode(input).replace(' [PAD]','')}.")
      has_unknown.append(i)
      break

# Checking the originals of those reviews
for index in has_unknown:
  print(X_test.iloc[index]['review'])

# Sample from above parenthetical)
untokenizable_sample = 'mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm'
len(untokenizable_sample)




It looks like the primary source of unknown tokens are excessively repeated characters, such as in "mmmmmmmmmm..." with 157 m's.

type(test_inputs)

X_test.columns

# Create a dask dataframe of the tokenized reviews
test_input_ddf = dd.from_dict({'index': X_test.index,
                               'input_ids': test_inputs['input_ids'],
                               'attention_mask': test_inputs['attention_mask']},
                              npartitions=10)

# Check that inputs match up
test_input_ddf.head(5)


# Check that the first shuffled index decodes as expected
tokenizer.decode(test_input_ddf.loc[test_input_ddf['index']==886662,'input_ids'].compute()[0])

X_test.head(1)

# View sample of attention_mask feature
test_input_ddf.head(1).attention_mask.values

# View sample of an input_ids feature
test_input_ddf.head(1).input_ids.values

# Sample tokens with their corresponding decoded values
tokenizer.decode(2023)

tokenizer.decode(1024)

len(test_input_ddf.head(1).attention_mask.values[0])

schema = pa.schema([('index', pa.int64()),
                    ('input_ids', pa.list_(pa.int64())),
                    ('attention_mask', pa.list_(pa.int64()))])

test_input_ddf.to_parquet('drive/MyDrive/BERT Sentiment/CSVs/test_inputs.parquet',schema=schema)


# X_train is too big to hold in memory. Going to need to batch them.
batch_size = len(X_train)//15 + 1
batch_size

# Tokenize reviews in batches
for i in range(15):
  start = i*batch_size
  end = min((i+1)*batch_size,len(X_train))
  train_inputs = tokenizer(list(X_train['review'].iloc[start:end]),truncation=True,padding=True)
  train_indices = X_train.index[start:end]
  train_input_ddf = dd.from_dict({'index': train_indices,
                                  'input_ids': train_inputs['input_ids'],
                                  'attention_mask': train_inputs['attention_mask']},
                                 npartitions=10)
  train_input_ddf.to_parquet(f'drive/MyDrive/BERT Sentiment/CSVs/train_inputs_{i}.parquet',schema=schema)


train_input_ddf.head(5)

# Test that the dataframes are correctly labeled.
train_inputs = tokenizer(list(X_train['review'].iloc[batch_size:2*batch_size]),truncation=True,padding=True)
train_indices = X_train.index[batch_size:2*batch_size]
train_input_ddf = dd.from_dict({'index': train_indices,
                                  'input_ids': train_inputs['input_ids'],
                                  'attention_mask': train_inputs['attention_mask']},
                                 npartitions=10)

train_input_ddf.head()

tokenizer.decode(train_input_ddf.loc[train_input_ddf['index']==2025630,'input_ids'].compute()[0])

X_train.iloc[batch_size]
