==> Notebooks/4_Hyperparameter_Tuning.ipynb <==
!pip install transformers[torch]
!pip install optuna

from google.colab import drive
drive.mount('/content/drive/')

import pandas as pd
from transformers import DistilBertForSequenceClassification, TrainingArguments, Trainer
from sklearn.metrics import precision_recall_fscore_support
import torch
import optuna

# Use 1/15 of training data for hyperparameter tuning
dev_set = pd.read_parquet('/content/drive/MyDrive/BERT Sentiment/CSVs/train_inputs_0.parquet')


y_train = pd.read_csv('/content/drive/MyDrive/BERT Sentiment/CSVs/y_train_full.csv')

# Double check that the two datasets match before combining them
dev_set.head()

y_train.head()

len(dev_set)

len(y_train)

dev_set.tail()

y_train.iloc[len(dev_set)-5:len(dev_set)]


dev_set = dev_set.join(y_train, how='inner')
len(dev_set)

# Inspect combined dataset
dev_set.tail()

# Drop duplicate column and set index
dev_set = dev_set.drop(columns=['Unnamed: 0','index'])


# Get dev set into expected format for model
dev_set.columns = ['input_ids','attention_mask','labels']
# Binary classification expects 0 and 1, not 1 and 2
dev_set['labels'] = dev_set['labels']-1
dev_set.head()

eval_set = dev_set.iloc[:len(dev_set)//10].reset_index(drop=True)
train_set = dev_set.iloc[len(dev_set)//10:].reset_index(drop=True)
train_set.head()



# Use a more reasonably-sized fine-tuning dataset
eval_set = eval_set.iloc[:1500]
train_set = train_set.iloc[:10000]


eval_set.head()

def model_init(trial):
      # Define hyperparameters
    learning_rate = trial.suggest_float("learning_rate", 1e-5, 5e-5, log=True)
    num_train_epochs = trial.suggest_int("num_train_epochs", 1, 3)
    gradient_accumulation_steps = trial.suggest_int("gradient_accumulation_steps", 1, 8)
    per_device_train_batch_size = trial.suggest_int("per_device_train_batch_size", 4, 16)
    evaluation_strategy = trial.suggest_categorical("evaluation_strategy", ['steps', 'epoch'])
    per_device_eval_batch_size = trial.suggest_int("per_device_eval_batch_size", 4, 16)
    warmup_steps = trial.suggest_int("warmup_steps", 100, 500)
    weight_decay = trial.suggest_float("weight_decay", 0.0, 0.1)

    model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased',num_labels=2)

    return model




class SentimentDataset(torch.utils.data.Dataset):
  def __init__(self, encodings, labels):
        self.encodings = encodings
        self.labels = labels

  def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        item['labels'] = torch.tensor(self.labels[idx])
        return item

  def __len__(self):
        return len(self.labels)

train_dataset = SentimentDataset(train_set, train_set['labels'])
eval_dataset = SentimentDataset(eval_set, eval_set['labels'])


len(train_dataset)

def objective(trial):

    # Define training arguments
    training_args = TrainingArguments(
        output_dir='drive/MyDrive/BERT Sentiment/output',
        logging_dir='drive/MyDrive/BERT Sentiment/output/logs',
        logging_steps=1000
    )

    # Initialize model
    model = model_init(trial)

    trainer = Trainer(
        model=model,
        args=training_args,
        train_dataset=train_dataset,
        eval_dataset=eval_dataset)

    trainer.train()


    # Evaluate model
    predictions = trainer.predict(eval_dataset)
    true_labels = eval_dataset.labels
    predicted_labels = predictions.predictions.argmax(-1)
    precision, recall, f1, _ = precision_recall_fscore_support(true_labels, predicted_labels, average='binary')
    print("Precision: " + str(precision))
    print("Recall: " + str(recall))
    print("F1: " + str(f1))

    return f1

study = optuna.create_study(direction='maximize')

study.optimize(objective, n_trials=8,n_jobs=6)

best_hyperparameters = study.best_params

print("Best hyperparameters" + str(best_hyperparameters))

print(study.best_trial)


import pickle

with open('/content/drive/MyDrive/BERT Sentiment/output/best_hyperparameters.pkl', 'wb') as outfile:
    pickle.dump(best_hyperparameters, outfile)
