==> Notebooks/5_Model_Creation.ipynb <==
from google.colab import drive
drive.mount('/content/drive/')

!pip install transformers[torch]
!pip install accelerate
!pip install evaluate

import pandas as pd
from transformers import Trainer, get_scheduler, TrainingArguments, AutoModelForSequenceClassification
from accelerate import Accelerator
import pickle
from torch.optim import AdamW
import torch
import numpy as np
import evaluate
from torch.utils.data import DataLoader
from tqdm.auto import tqdm
from sklearn.metrics import classification_report




# Load tuned hyperparameters
with open('/content/drive/MyDrive/BERT Sentiment/output/best_hyperparameters.pkl', 'rb') as infile:
    best_hyperparameters = pickle.load(infile)


best_hyperparameters


# Create Distilbert Model
model = AutoModelForSequenceClassification.from_pretrained(
    "distilbert-base-uncased",
    num_labels=2)


# Set training arguments
training_args = TrainingArguments(learning_rate = best_hyperparameters['learning_rate'],
                                  num_train_epochs = best_hyperparameters['num_train_epochs'],
                                  per_device_train_batch_size = best_hyperparameters['per_device_train_batch_size'],
                                  per_device_eval_batch_size = best_hyperparameters['per_device_eval_batch_size'],
                                  gradient_accumulation_steps = best_hyperparameters['gradient_accumulation_steps'],
                                  weight_decay = best_hyperparameters['weight_decay'],
                                  warmup_steps = best_hyperparameters['warmup_steps'],
                                  evaluation_strategy = best_hyperparameters['evaluation_strategy'],
                                  output_dir='drive/MyDrive/BERT Sentiment/output',
                                  logging_dir='drive/MyDrive/BERT Sentiment/output/logs',
                                  logging_steps=1000)


# Create evaluation metric
metric = evaluate.load('f1')

def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    return metric.compute(predictions=predictions, references=labels)



# Load about a million datapoints as the train dataset
train_sets = [pd.read_parquet('/content/drive/MyDrive/BERT Sentiment/CSVs/train_inputs_1.parquet'),
              pd.read_parquet('/content/drive/MyDrive/BERT Sentiment/CSVs/train_inputs_2.parquet'),
              pd.read_parquet('/content/drive/MyDrive/BERT Sentiment/CSVs/train_inputs_3.parquet')]





# Reduce training size
train_set = pd.concat(train_sets)[:500000]

# Clear duplicate from memory
del(train_sets)

# Choose more training data to be eval set
eval_set = pd.read_parquet('/content/drive/MyDrive/BERT Sentiment/CSVs/train_inputs_5.parquet')


train_set.head()

# Retrieve labels for train/eval data
y_train = pd.read_csv('/content/drive/MyDrive/BERT Sentiment/CSVs/y_train_full.csv')


y_train.columns = ['index','labels']

y_train.head()

# Append labels to datasets
train_set = pd.merge(train_set, y_train, on='index')
eval_set = pd.merge(eval_set, y_train, on='index')


eval_set.head()

# Reformat labels from 1 and 2 to 0 and 1
train_set['labels'] = train_set['labels'] - 1
eval_set['labels'] = eval_set['labels'] - 1

# Format dataframe for use in model
train_set.drop(columns=['index'], inplace=True)
eval_set.drop(columns=['index'], inplace=True)

# Convert dataframes to torch dataset objects
class SentimentDataset(torch.utils.data.Dataset):
  def __init__(self, encodings, labels):
        self.encodings = encodings
        self.labels = labels

  def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        item['labels'] = torch.tensor(self.labels[idx])
        return item

  def __len__(self):
        return len(self.labels)

train_dataset = SentimentDataset(train_set, train_set['labels'])
eval_dataset = SentimentDataset(eval_set, eval_set['labels'])

# Create objects for parallel training
accelerator = Accelerator()

train_dataloader = DataLoader(train_dataset, batch_size=48, shuffle=True)
eval_dataloader = DataLoader(eval_dataset, batch_size=48, shuffle=True)


optimizer = AdamW(model.parameters(), lr=best_hyperparameters['learning_rate'])

train_dataloader, eval_dataloader, model, optimizer = accelerator.prepare(train_dataloader, eval_dataloader, model, optimizer)


trainer = Trainer(model=model, args=training_args, train_dataset=train_dataset, eval_dataset=eval_dataset, compute_metrics=compute_metrics)

num_epochs = 3
num_training_steps = num_epochs * len(train_dataloader)
lr_scheduler = get_scheduler(
    "linear",
    optimizer=optimizer,
    num_warmup_steps=0,
    num_training_steps=num_training_steps
)


progress_bar = tqdm(range(num_training_steps))

# Train model
model.train()
for epoch in range(num_epochs):
  for batch in train_dataloader:
    outputs = model(**batch)
    loss = outputs.loss
    accelerator.backward(loss)
    optimizer.step()
    lr_scheduler.step()
    optimizer.zero_grad()
    progress_bar.update(1)


# Save trained model
torch.save(model.state_dict(), 'drive/MyDrive/BERT Sentiment/Models/sentiment_model.pt')


# Test model on unseen dataset
evaluation_set = pd.read_parquet('/content/drive/MyDrive/BERT Sentiment/CSVs/train_inputs_4.parquet')


y_train = pd.read_csv('/content/drive/MyDrive/BERT Sentiment/CSVs/y_train_full.csv')
y_train.columns = ['index','labels']
y_train['labels'] = y_train['labels'] - 1


y_train.head()

evaluation_set = evaluation_set.merge(y_train, on='index').reset_index(drop=True)

evaluation_set.head()

evaluation_dataset = SentimentDataset(evaluation_set, evaluation_set['labels'])


model.eval()

total_loss = 0
y_pred = []
y_true = []

with torch.no_grad():
  for batch in tqdm(evaluation_dataset):
    input_ids = batch['input_ids'].to('cuda:0')
    attention_mask = batch['attention_mask'].to('cuda:0')
    labels = batch['labels'].to('cuda:0')

    model_outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
    loss = model_outputs.loss
    total_loss += loss.item()

    logits = model_outputs.logits
    predictions = torch.argmax(logits, dim=-1)
    y_pred.extend(predictions)
    y_true.extend(batch['labels'].to('cuda:0'))






total_loss

len(y_pred)

y_true = []
with torch.no_grad():
  for batch in tqdm(evaluation_dataset):
    y_true.append(batch['labels'].to('cuda:0'))



# Turn tensors to ints
for i in range(len(y_true)):
  y_true[i] = y_true[i].item()
  y_pred[i] = y_pred[i].item()



print(classification_report(y_true, y_pred))
